Handle checkstyle linewith check as seperate module
The checkstyle linewidth check was based on the java parser without need.
See https://github.com/checkstyle/checkstyle/issues/2116 for details.

Newer versions of checkstyle rip out the linewidth check out of the
treewalker module an put it into a seperate module. Use this seperate
module for the linewidth check.

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>